added functionality to the projects and filetree to distinguish between files
